refactor(cli): consolidate 14 command files into 7, delete runtime/ folder
Co-locate related commands to reduce file surface area and indirection:
- data.ts: merge get, list, count, delete, tables, export (table CRUD)
- project.ts: merge init, install, uninstall (project management)
- start.ts: merge start-command + start-daemon (was split across 2 folders)
- kv.ts, auth.ts, run.ts, describe.ts: rename from *-command.ts

Create util/workspace.ts merging:
- runtime/open-workspace.ts (one-shot lifecycle)
- util/with-workspace-options.ts (yargs options, runCommand, resolveTable)
- util/paths.ts (resolveEpicenterHome inlined into cli.ts)

Introduce universal runCommand() replacing duplicated try/catch in 6 files.
Linearize install handler: extract installWorkspace() with inline guards
instead of 8 early-return outputError blocks.

Delete: 14 old command files, runtime/ folder, with-workspace-options.ts,
paths.ts. 14 command files → 7. 4 folders → 3.
