+++ packages/cli/src/index.ts
@@ -1,6 +1,5 @@
 /** @module @epicenter/cli — Public API for the Epicenter CLI package. */
 
-export { createCLI } from './cli';
+export { createCLI, resolveEpicenterHome } from './cli';
 export { createAuthApi, type AuthApi } from './auth/api';
-export { resolveEpicenterHome } from './util/paths';
 export { loadConfig, type LoadConfigResult } from './load-config';
